Changed handlers mutability.
In the real world streaming parsers must be mutable.

Now it's allowed to change its internal state to be able to track parsing process, cause the parser itself is streaming. For example now it's possible to collect partial headers/body chunks inside the handler for further (or immediate) aggregating or processing.